+++ pkg/analysis_server/test/constants.dart
@@ -2,31 +2,30 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-const String CODE = 'code';
-const String COMPLETION_RESULTS = 'completion.results';
-const String CONTEXT_MESSAGES = 'contextMessages';
-const String CORRECTION = 'correction';
+const String codeKey = 'code';
+const String contextReferencesKey = 'contextMessages';
+const String correctionKey = 'correction';
 
 /// Useful for debugging locally, setting this to true will cause all JSON
 /// communication to be printed to stdout.
 const debugPrintCommunication = false;
 
-const String EDITS = 'edits';
-const String END_COLUMN = 'endColumn';
-const String END_LINE = 'endLine';
-const String FILE = 'file';
-const String FILE_STAMP = 'fileStamp';
-const String HAS_FIX = 'hasFix';
-const String LENGTH = 'length';
-const String LOCATION = 'location';
-const String MESSAGE = 'message';
-const String OFFSET = 'offset';
-const String REPLACEMENT = 'replacement';
-const String SEVERITY = 'severity';
-const String STACK_TRACE = 'stackTrace';
-const String START_COLUMN = 'startColumn';
-const String START_LINE = 'startLine';
-const String SUBSCRIPTIONS = 'subscriptions';
-const String TYPE = 'type';
-const String URL = 'url';
-const String VERSION = 'version';
+const String editsKey = 'edits';
+const String endColumnKey = 'endColumn';
+const String endLineKey = 'endLine';
+const String fileKey = 'file';
+const String fileStampKey = 'fileStamp';
+const String hasFixKey = 'hasFix';
+const String lengthKey = 'length';
+const String locationKey = 'location';
+const String messageKey = 'message';
+const String offsetKey = 'offset';
+const String replacementKey = 'replacement';
+const String severityKey = 'severity';
+const String stackTraceKey = 'stackTrace';
+const String startColumnKey = 'startColumn';
+const String startLineKey = 'startLine';
+const String subscriptionsKey = 'subscriptions';
+const String typeKey = 'type';
+const String urlKey = 'url';
+const String versionKey = 'version';