+++ pkg/analysis_server/test/domain_server_test.dart
@@ -77,7 +77,7 @@ class ServerDomainTest extends PubPackageAnalysisServerTest {
       response.toJson(),
       equals({
         Response.idAttributeName: '0',
-        Response.resultAttributeName: {VERSION: PROTOCOL_VERSION},
+        Response.resultAttributeName: {versionKey: PROTOCOL_VERSION},
       }),
     );
   }
@@ -353,7 +353,7 @@ class ServerDomainTest extends PubPackageAnalysisServerTest {
 
   Future<void> test_setSubscriptions_invalidServiceName() async {
     var request = Request('0', SERVER_REQUEST_SET_SUBSCRIPTIONS, {
-      SUBSCRIPTIONS: ['noSuchService'],
+      subscriptionsKey: ['noSuchService'],
     });
     var response = await handleRequest(request);
     expect(response, isResponseFailure('0'));